+++ src_cpp/raycast.cpp
@@ -242,8 +242,8 @@ InterestPoint* get_interest_points( const Matrix& J_norm, const Point& earth_pos
                                     double theta_min, double theta_max, 
                                     int nb_theta, int nb_phi, 
                                     double dx, double dr,
-                                    double alpha_0_min, double alpha_0_max, double nb_alpha_0,
-                                    double r_0_mult_min, double r_0_mult_max, double nb_r_0,
+                                    double alpha_0_min, double alpha_0_max, int nb_alpha_0,
+                                    double r_0_mult_min, double r_0_mult_max, int nb_r_0,
                                     double* p_avg_std_dev )
 {
     std::vector<double>* interest_radii_candidates = new std::vector<double>[ nb_theta*nb_phi ];
@@ -322,8 +322,8 @@ InterestPoint* get_interest_points( const Matrix& J_norm, const Point& earth_pos
                                     double theta_min, double theta_max, 
                                     int nb_theta, int nb_phi, 
                                     double dx, double dr,
-                                    double alpha_0_min, double alpha_0_max, double nb_alpha_0,
-                                    double r_0_mult_min, double r_0_mult_max, double nb_r_0,
+                                    double alpha_0_min, double alpha_0_max, int nb_alpha_0,
+                                    double r_0_mult_min, double r_0_mult_max, int nb_r_0,
                                     double* p_avg_std_dev )
 {
     std::vector<Point> bow_shock = get_bowshock( Rho, earth_pos, dr, nb_phi, nb_theta, false );
